+++ es-ES/index.js
@@ -1,0 +1,64 @@
+import common from './common'
+import form from './form'
+import emails from './emails'
+import blog from './blog'
+import challenges from './challenges'
+import discovery from './discovery'
+import encyclopedia from './encyclopedia'
+import enrollment from './enrollment'
+import feedback from './feedback'
+import feed from './feed'
+import genre from './genre'
+import groups from './groups'
+import licenses from './licenses'
+import organizations from './organizations'
+import pages from './pages'
+import plans from './plans'
+import pm from './pm'
+import rating from './rating'
+import settings from './settings'
+import sharing from './sharing'
+import stats from './stats'
+import workshop from './workshop'
+import comments from './comments'
+import library from './library'
+import legal from './legal'
+import reader from './reader'
+import writer from './writer'
+import universe from './universe'
+import audit from './audit'
+import user from './user'
+
+export default {
+  locales: ['es', 'es-ES', 'es-AR', 'es-BO', 'es-BR', 'es-BZ', 'es-CL', 'es-CO', 'es-CR', 'es-CU', 'es-DO', 'es-EA', 'es-EC', 'ES-GQ', 'es-GT', 'es-HN', 'es-IC', 'es-MX', 'es-NI', ' es-PA', 'es-PE', 'es-PH', 'es-PR', 'es-PY', 'es-SV', 'es-US', 'es-UY', 'es-VE'],
+  ...common,
+  ...form,
+  ...emails,
+  ...blog,
+  ...challenges,
+  ...discovery,
+  ...encyclopedia,
+  ...enrollment,
+  ...feedback,
+  ...feed,
+  ...genre,
+  ...groups,
+  ...licenses,
+  ...organizations,
+  ...pages,
+  ...plans,
+  ...pm,
+  ...rating,
+  ...settings,
+  ...sharing,
+  ...stats,
+  ...workshop,
+  ...comments,
+  ...library,
+  ...writer,
+  ...reader,
+  ...legal,
+  ...universe,
+  ...audit,
+  ...user
+}